refactor(mcp-http): fix layer inversion — move session factories into mcp/ layer
createStreamableHttpHandler and createSseHandlers belonged in the mcp/ layer
(MCP transport concerns), but lived in server/mcp-http.ts, which caused
mcp/http-transport.ts to import from server/ — inverting the established
server/ → mcp/ dependency direction.

Move both factories (plus MCPSession/SSESession interfaces and TTL constants)
into mcp/http-transport.ts. server/mcp-http.ts now only contains
mountMCPEndpoints, which imports from mcp/http-transport.ts — restoring the
correct arrow direction with no functional change.

Update test imports accordingly (createStreamableHttpHandler/createSseHandlers
now come from mcp/http-transport.js; mountMCPEndpoints still from server/mcp-http.js).
